refactor: mutualize TypeScript and JavaScript module resolvers using the ECMAScript standard
