+++ packages/skott/src/modules/walkers/ecmascript/index.ts
@@ -1,0 +1,2 @@
+export { JavaScriptModuleWalker } from "./javascript/walker.js";
+export { TypeScriptModuleWalker } from "./typescript/walker.js";
