+++ packages/skott/src/modules/walkers/ecmascript/javascript/cjs.ts
