+++ packages/skott/src/modules/walkers/ecmascript/module-declaration.ts
